feat: :sparkles: Adicionando novos gráficos e melhorando a fct tpv metodo pagamento
